+++ examples/vue-composable-example/src/main.js
@@ -1,6 +1,7 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
 import VueCompositionApi from "@vue/composition-api";
+import { hydrationPlugin } from "vue-composable";
 import App from "./App.vue";
 
 import HelloWorld from "./components/HelloWorld";
@@ -13,6 +14,8 @@ import SharedRef from "./components/SharedRef";
 
 Vue.use(VueRouter);
 Vue.use(VueCompositionApi);
+Vue.use(hydrationPlugin);
+
 window.SuperVue = Vue;
 
 Vue.config.productionTip = false;