+++ packages/vue-composable/src/misc/index.ts
@@ -1,2 +1,4 @@
 export * from "./matchMedia";
 export * from "./sharedRef";
+export * from "./vmodel";
+export * from "./injectFactory";